requirements.lock.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
datasets==3.1.0
deepspeed==0.15.4
dill==0.3.8
einops==0.7.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.8
hjson==3.1.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
moshi==0.1.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
networkx==3.4.2
ninja==1.13.0
numpy==2.1.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.4.5
sentencepiece==0.2.0
sentry-sdk==2.35.1
six==1.17.0
smmap==5.0.2
sounddevice==0.5.0
soundfile==0.13.1
sphn==0.2.0
sympy==1.14.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.4.1
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.4
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.1
xxhash==3.5.0
yarl==1.20.1
